TTT_LORA_BIAS: cast bias to out.dtype in forward to preserve bf16
Bug: nn.Parameter(torch.zeros(bsz, 1, out_features)) defaults to float32.
Under autocast, (bf16 @ fp32) @ fp32 → bf16 output, but `out + self.bias`
type-promotes to fp32, which then flows into attention. flash-attn only
accepts fp16/bf16/fp8, so mha_fwd raises RuntimeError.

Fix: out = out + self.bias.to(out.dtype), same pattern used elsewhere
in the file for bias additions under autocast.

+++ records/track_10min_16mb/2026-04-19_SP8192_CaseOps_GatedAttn_QuantGate_Loop45_PhasedTTT/train_gpt.py
@@ -1417,7 +1417,9 @@ class BatchedLinearLoRA(nn.Module):
     def forward(self, x):
         out = (x @ self.A.transpose(1, 2)) @ self.B.transpose(1, 2)
         if self.bias is not None:
-            out = out + self.bias
+            # Cast bias to out.dtype so it doesn't upcast (bf16 + fp32 → fp32)
+            # and break flash-attn downstream (which only accepts fp16/bf16/fp8).
+            out = out + self.bias.to(out.dtype)
         return out
